Start splitting completion into components
